server/pyproject.toml
[tool.poetry]
name = "ChatCLI"
version = "0.1.0"
description = ""
authors = ["ryoya0902 <ryuasa1192@gmail.com>"]
readme = "README.md"
packages = [{include = "ChatCLI"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.3"
isort = "^5.12.0"
black = "^23.3.0"
openai = "^0.27.8"
fastapi = "^0.98.0"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"
pyproject-flake8 = "^6.0.0.post1"
isort = "^5.12.0"
flake8-isort = "^6.0.0"
flake8-bugbear = "^23.5.9"
flake8-builtins = "^2.1.0"
flake8-unused-arguments = "^0.0.13"
flake8-pytest-style = "^1.7.2"
pep8-naming = "^0.13.3"
mypy = "^1.3.0"

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.black]
line-length = 128

[tool.isort]
profile = "black"
line_length = 128
multi_line_output = 3

[tool.flake8]
max-line-length = 128
extend-ignore = ["E203", "W503"]
exclude = [".venv", ".git", "__pycache__"]
max-complexity = 10